gw: add tile id mercator

+++ backend/src/main/kotlin/ru/dmitry4k/geomarkback/service/dto/TileId.kt
@@ -1,7 +1,7 @@
 package ru.dmitry4k.geomarkback.service.dto
 
 data class TileId(
-    val x: Int,
-    val y: Int,
+    val x: Double,
+    val y: Double,
     val z: Int
 )
